app/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
pydantic[email]==2.12.5
python-jose[cryptography]>=3.3.0
bcrypt>=5.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
